receipt logic fixed
